% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Ms.R
\name{calculate_Ms}
\alias{calculate_Ms}
\title{Calculates var\link{M(Xs)}}
\usage{
calculate_Ms(model_data)
}
\arguments{
\item{model_data}{model_data object that ships with the package}
}
\value{
matrix
}
\description{
Calculates var\link{M(Xs)}
}
