% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sst.R
\name{calculate_sst_update}
\alias{calculate_sst_update}
\title{Calculate the SST update}
\usage{
calculate_sst_update(
  simulation,
  proxy,
  H_list,
  params,
  incl_varM = FALSE,
  sst_min = -1.92
)
}
\arguments{
\item{simulation}{A sst_sim object containing the sst simulations.
Such an object can be created using the create_sst_sim_object() command.}

\item{proxy}{A sst_data object containing the proxy data. Such and object
can be created using the create_sst_data_object() command.}

\item{H_list}{the output from generate_Hx() containing a list of the
incidence matrices.}

\item{params}{a tibble of parameters for the analysis with columns tau, c
kappa and alpha. These parameters are defined in Section 5.2 of the paper.}

\item{incl_varM}{include the varM matrix in the calculation.}

\item{sst_min}{set minimum value of SST. The default value is the
scientifically understood value. We scale MME outputs to this value (in
practice this makes miniscule difference to the analysis).}
}
\value{
Returns a sst_reconstruction object.
}
\description{
This requires the simulations, the data, the incidence matrices, and the
analysis parameters
}
