% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_reconstruction_class.R
\name{calculate_svds.sst_reconstruction}
\alias{calculate_svds.sst_reconstruction}
\title{Calculates the top n SVDs from an sst_reconstruction object}
\usage{
\method{calculate_svds}{sst_reconstruction}(obj, n = 20)
}
\arguments{
\item{obj}{an sst_reconstruction object}

\item{n}{the number of SVDs}
}
\value{
a list comprising of a tibble with the sampled field and the SVD
indexes
}
\description{
Calculates the top n SVDs from an sst_reconstruction object
}
