% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{coastlines}
\alias{coastlines}
\title{Coastline shapefile}
\format{
a shapefile
}
\usage{
coastlines
}
\description{
Shapefile of global coastlines for plotting.
}
\keyword{datasets}
