% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{coords_area}
\alias{coords_area}
\title{Grid-cell areas}
\format{
a tibble
}
\usage{
coords_area
}
\description{
Grid-cell areas for the PMIP ocean grid. Used in calculation of sea-ice areas
}
\keyword{datasets}
