% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_basis.R
\name{create_psii}
\alias{create_psii}
\title{Calculate Psi Matrix}
\usage{
create_psii(model_tmp, spline_params)
}
\arguments{
\item{model_tmp}{a tibble of MME member SST and SIC}

\item{spline_params}{prior spline specifications
w}
}
\value{
returns Matrix object
}
\description{
Calculate Psi Matrix
}
