% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_class.R
\name{create_sst_sim_object}
\alias{create_sst_sim_object}
\title{Crate a sst_sim object}
\usage{
create_sst_sim_object(data, means, coords)
}
\arguments{
\item{object}{List with elements XYZ}
}
\value{
Returns a sst_sim object
}
\description{
This creates the object that holds the sst_sim results, i.e., the X
}
