% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_basis.R
\name{create_theta}
\alias{create_theta}
\title{Calculate Theta Matrix}
\usage{
create_theta(all_fits, idx_select)
}
\arguments{
\item{all_fits}{}

\item{idx_select}{}
}
\value{
returns Matrix object
}
\description{
Calculate Theta Matrix
}
