% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_spline}
\alias{fit_spline}
\title{Calculates the I-spline design matrix from a design}
\usage{
fit_spline(sst, ice, spline_params)
}
\arguments{
\item{x_in}{the inputs at which the basis is to be evaluated}

\item{boundary}{the boundary points}

\item{knots}{the knot locations}

\item{degree}{the spline degree}
}
\value{
a spline basis design matrix
}
\description{
Calculates the I-spline design matrix from a design
}
