% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_H.R
\name{generate_H}
\alias{generate_H}
\title{Generates the spatial and temporal incidence matrices}
\usage{
generate_H(simulation, proxy)
}
\arguments{
\item{simulation}{A sst_sim object containing the sst simulations}

\item{proxy}{A sst_data object containing the proxy data}
}
\value{
Returns a list of Hs, Hs_collapse (reshuffled Hs to make some
calculation easier), Ht and H.
}
\description{
Generates the spatial and temporal incidence matrices
}
