% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Hy.R
\name{generate_Hy}
\alias{generate_Hy}
\title{Generates the SIC incidence matrix}
\usage{
generate_Hy(coords, ns)
}
\arguments{
\item{coords}{A sst_sim object containing the sst simulations}

\item{ns}{A sst_data object containing the proxy data}
}
\value{
Returns a matrix
}
\description{
Generates the SIC incidence matrix
}
