% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nroy.R
\name{is_nroy}
\alias{is_nroy}
\title{Tests NROY membership}
\usage{
is_nroy(sampled_ice, wsi_area_south, wsi_area_north, coords_area)
}
\arguments{
\item{sampled_ice}{output of sic field from sample_fields()}

\item{wsi_area_south}{mean Southern winter sea ice area}

\item{wsi_area_north}{mean Northern winter sea ice area}

\item{coords_area}{grid-cell areas. This tibble should ship with the package}
}
\value{
Either TRUE or FALSE
}
\description{
A history matching helper function that assesses if a plausibility sample is
part of the NROY space of sea-ice concentrations
}
