% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{margo_sst}
\alias{margo_sst}
\title{SST Proxy Data}
\format{
A sst_data object of PMIP outputs
}
\usage{
margo_sst
}
\description{
A collection of proxy based SST measurements from the MARGO project.
}
\keyword{datasets}
