% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{model_data}
\alias{model_data}
\title{Monthly averages of SST and SIC from PMIP}
\format{
A sst_sim object of PMIP outputs
}
\usage{
model_data
}
\description{
A collection of SST and SIC Simulations from PMIP 3 and 4.
}
\keyword{datasets}
