% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_data_class.R
\name{new_sst_data}
\alias{new_sst_data}
\title{Helper function to assign class to a sst_data object}
\usage{
new_sst_data(object)
}
\arguments{
\item{object}{sst_data object from sst_data()}
}
\value{
Returns a sst_data object
}
\description{
Helper function to assign class to a sst_data object
}
