% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_reconstruction_class.R
\name{new_sst_reconstruction}
\alias{new_sst_reconstruction}
\title{Helper function to assign class to a sst_reconstruction object}
\usage{
new_sst_reconstruction(object)
}
\arguments{
\item{object}{sst_reconstruction object from sst_reconstruction()}
}
\value{
Returns a sst_reconstruction object
}
\description{
Helper function to assign class to a sst_reconstruction object
}
