% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_class.R
\name{new_sst_sim}
\alias{new_sst_sim}
\title{Helper function to assign class to a sst_sim object}
\usage{
new_sst_sim(object)
}
\arguments{
\item{object}{sst_sim object from sst_sim()}
}
\value{
Returns a sst_sim object
}
\description{
Helper function to assign class to a sst_sim object
}
