% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sst_parameters.R
\name{parameter_loss}
\alias{parameter_loss}
\title{Defines the loss function with which to train the sst parameters}
\usage{
parameter_loss(params, error, D, varB, noise, Hs)
}
\arguments{
\item{params}{C4-Wendland parameters}

\item{error}{error vector}

\item{D}{data}

\item{varB}{bias variance}

\item{noise}{observation error}

\item{Hs}{spatial incidence matrix}
}
\value{
Returns loss
}
\description{
This is currently negative log-likelihood, investigate using bias-variance
}
