% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_spline_basis}
\alias{plot_spline_basis}
\title{Plots a chosen spline basis}
\usage{
plot_spline_basis(spline_object)
}
\arguments{
\item{spline_object}{Output from calc_X}
}
\value{
A plot of the spline bases
}
\description{
Plots a chosen spline basis
}
