% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pmip_sst}
\alias{pmip_sst}
\title{PMIP SST Simulations}
\format{
A sst_sim object of PMIP outputs
}
\source{
Go through and ref every object in here
}
\usage{
pmip_sst
}
\description{
A collection of SST Simulations from PMIP 3 and 4.
}
\keyword{datasets}
