% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_data_class.R
\name{print.sst_data}
\alias{print.sst_data}
\title{Print function for sst_data object}
\usage{
\method{print}{sst_data}(object)
}
\arguments{
\item{object}{sst_data object}
}
\value{
Prints sst_data object
}
\description{
Print function for sst_data object
}
