% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_reconstruction_class.R
\name{print.sst_reconstruction}
\alias{print.sst_reconstruction}
\title{Print function for sst_reconstruction object}
\usage{
\method{print}{sst_reconstruction}(x, ...)
}
\arguments{
\item{x}{sst_reconstruction object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints sst_reconstruction object
}
\description{
Print function for sst_reconstruction object
}
