% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_betais.R
\name{project_betais}
\alias{project_betais}
\title{Find the beta projection}
\usage{
project_betais(sst_update, spline_params)
}
\arguments{
\item{sst_update}{a sst_sim object containing the sst simulations}

\item{spline_params}{prior spline specifications}
}
\value{
list of betais
}
\description{
Find the beta projection
}
