% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ind_coefficients.R
\name{project_sic_coefs}
\alias{project_sic_coefs}
\title{Projects sst/sic data at each location onto the spline basis coefficients}
\usage{
project_sic_coefs(sst_update, spline_params)
}
\arguments{
\item{sst_update}{an updated sst object from calculate_sst_update()}

\item{spline_params}{chosen spline parameters. This is a list with elements
bounds, knots, degree, and prior_exp}
}
\value{
a tibble of all of the spline basis coefficients at each location
}
\description{
Projects sst/sic data at each location onto the spline basis coefficients
}
