% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fields.R
\name{sample_fields}
\alias{sample_fields}
\title{Generates a sic field sample from sampled SVD indexes}
\usage{
sample_fields(sst_update, ice_updates, spline_params, svd_sample)
}
\arguments{
\item{sst_update}{an updated sst object from calculate_sst_update()}

\item{ice_updates}{an updated sic object from update_sic()}

\item{spline_params}{chosen spline parameters. This is a list with elements
bounds, knots, degree, and prior_exp}

\item{svd_sample}{sampled indexes of the sst SVDs}
}
\value{
a list comprising of a tibble with the sampled field and the SVD
indexes
}
\description{
Generates a sic field sample from sampled SVD indexes
}
