% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sic_data}
\alias{sic_data}
\title{SIC Data}
\format{
A sst_sim object of PMIP outputs
}
\usage{
sic_data
}
\description{
binary estimates of SIC inferred from Northern and Southern hemisphere maximum sea-ice extents
}
\keyword{datasets}
