% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sst_data}
\alias{sst_data}
\title{SST Data}
\format{
An object of class \code{sst_data} of length 5.
}
\usage{
sst_data
}
\description{
Proxy-reconstructions of paleo-SST data, predominantly from the MARGO project.
}
\keyword{datasets}
