% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_reconstruction_class.R
\name{sst_reconstruction}
\alias{sst_reconstruction}
\title{Crate a sst_reconstruction object}
\usage{
sst_reconstruction(object)
}
\arguments{
\item{object}{List with elements XYZ}
}
\value{
Returns a sst_reconstruction object
}
\description{
This creates the object that holds the sst_reconstruction results, i.e., the X
}
