% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_class.R
\name{sst_sim}
\alias{sst_sim}
\title{Crate a sst_sim object}
\usage{
sst_sim(object)
}
\arguments{
\item{object}{List with elements XYZ}
}
\value{
Returns a sst_sim object
}
\description{
This creates the object that holds the sst_sim results, i.e., the X
}
