% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sst_parameters.R
\name{train_params}
\alias{train_params}
\title{Crate a sst_data object}
\usage{
train_params(simulation, proxy, H_list)
}
\arguments{
\item{simulation}{a sst_sim object containing the sst simulations}

\item{proxy}{a sst_data object containing the proxy data}

\item{H_list}{the output from generate_H() containing a list of the
incidence matrices}
}
\value{
Returns a sst_data object
}
\description{
This creates the object that holds the sst_data results, i.e., the Z
}
