% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sic.R
\name{update_sic}
\alias{update_sic}
\title{Calculates BL update of the SIC parameters}
\usage{
update_sic(all_ice_fits, spline_params)
}
\arguments{
\item{all_ice_fits}{sic coefficient projections from project_sic_coefs()}

\item{spline_params}{chosen spline parameters. This is a list with elements
bounds, knots, degree, and prior_exp}
}
\value{
list with sic updates for each parameter
}
\description{
Calculates BL update of the SIC parameters
}
