% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wendland}
\alias{wendland}
\title{C4-Wendland covariance function}
\usage{
wendland(D, tau, alpha, kappa, delta = NULL)
}
\arguments{
\item{D}{Distance matrix}

\item{tau}{Length-scale}

\item{alpha}{Roughness}

\item{kappa}{Amplitude parameter}

\item{delta}{Nugget term}
}
\value{
Returns a covariance matrix
}
\description{
Calculates the C4-Wendland covariance matrix given an input distance matrix
}
