% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wsi}
\alias{wsi}
\title{Winter sea-ice extents}
\format{
a tibble
}
\source{
Gersonde, R., Crosta, X., Abelmann, A., & Armand, L. (2005).
Sea-surface temperature and sea ice distribution of the Southern Ocean at
the EPILOG Last Glacial Maximum—a circum-Antarctic view based on siliceous
microfossil records. Quaternary science reviews, 24(7-9), 869-896.
}
\usage{
wsi
}
\description{
Winter sea-ice extents for Antarctica from Gersonde (2005)
}
\keyword{datasets}
