% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wsi_extent}
\alias{wsi_extent}
\title{Winter sea-ice extents}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 364 rows and 3 columns.
}
\usage{
wsi_extent
}
\description{
Winter sea-ice extents
}
\keyword{datasets}
