% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sic.R
\name{calculate_sic_update}
\alias{calculate_sic_update}
\title{Calculates BL update of the SIC parameters}
\usage{
calculate_sic_update(
  sst_update,
  spline_params,
  sic_prior_params,
  betais,
  Xstar,
  sic_data
)
}
\arguments{
\item{spline_params}{chosen spline parameters. This is a list with elements
bounds, knots, degree, and prior_exp}

\item{all_ice_fits}{sic coefficient projections from project_sic_coefs()}
}
\value{
list with sic updates for each parameter
}
\description{
Calculates BL update of the SIC parameters
}
