% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sst.R
\name{calculate_sst_update}
\alias{calculate_sst_update}
\title{Calculate the SST update}
\usage{
calculate_sst_update(simulation, proxy, H_list, params, incl_varM = NULL)
}
\arguments{
\item{simulation}{a sst_sim object containing the sst simulations}

\item{proxy}{a sst_data object containing the proxy data}

\item{H_list}{the output from generate_H() containing a list of the
incidence matrices}

\item{params}{fit parameters from train_params()}

\item{incl_varM}{include the varM matrix in the calculation}
}
\value{
Returns loss
}
\description{
This requires the simulations, the data, the incidence matrices, and the
trained parameters
}
