% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_class.R
\name{create_sst_sim_object}
\alias{create_sst_sim_object}
\title{Crate a sst_sim object}
\usage{
create_sst_sim_object(data, means, coords)
}
\arguments{
\item{data}{a tibble containing the simulation data. Expected column names
are lat, lon, time, sst, ice and model. These are explained further
in the package README.}

\item{means}{the spatio-temporal SST means, i.e. averages over the MME for
each location in space and time.}

\item{coords}{the unique spatial lat/lon coords.}
}
\value{
Returns a sst_sim object
}
\description{
This creates the object that holds the sst_sim results, i.e., the X
}
