% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{margo_sst}
\alias{margo_sst}
\title{SST Proxy Data}
\format{
A sst_sim object of PMIP outputs
}
\source{
Go through and ref every object in here
}
\usage{
margo_sst
}
\description{
A collection of proxy based SST measurements
}
\keyword{datasets}
