% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pmip_sst}
\alias{pmip_sst}
\title{PMIP SST Simulations}
\format{
A sst_sim object of PMIP outputs
}
\usage{
pmip_sst
}
\description{
A collection of SST Simulations from PMIP 3 and 4. This list is not
exhaustive but is what was available to the project at the time of
publication.
}
\keyword{datasets}
