% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_class.R
\name{print.sst_sim}
\alias{print.sst_sim}
\title{Print function for sst_sim object}
\usage{
\method{print}{sst_sim}(object)
}
\arguments{
\item{object}{sst_sim object}
}
\value{
Prints sst_sim object
}
\description{
Print function for sst_sim object
}
