% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_betais.R
\name{project_betais}
\alias{project_betais}
\title{Find the beta projection}
\usage{
project_betais(sst_update, spline_params)
}
\arguments{
\item{sst_update}{an sst_reconstruction object generated from
calculate_sst_update(). This contains the results of the SST updates.}

\item{spline_params}{prior specifications of spline functions. A named-list
with names bounds, knots, degree and prior_exp (default coefficients).}
}
\value{
list containing betais, Theta and the mean of the betais.
}
\description{
Find the beta projection
}
