% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_data_class.R
\name{sst_data}
\alias{sst_data}
\title{Crate a sst_data object}
\usage{
sst_data(object)
}
\arguments{
\item{object}{List with elements XYZ}
}
\value{
Returns a sst_data object
}
\description{
This creates the object that holds the sst_data results, i.e., the Z
}
